dingers/requirements.txt
MLB-StatsAPI>=1.0.0
requests>=2.31.0
tenacity>=8.2.0
google-cloud-firestore>=2.16.0